+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [nickname] feature: now the `nickname format` supports the `placeholders`. (#438)
-- [chat.trigger] fix: use vanilla Java Matcher to handle the regex replacement, so it will be consistent for complex regex and replacement.
+- [chat.trigger] improvements and fixes
+  - fix: use vanilla Java Matcher to handle the regex replacement, so it will be consistent for complex regex and replacement.
+  - feature: pre-compile the pattern to improve the performance.
+- [chat.replace] improvements and fixes
+  - fix: use vanilla Java Matcher to handle the regex replacement, so it will be consistent for complex regex and replacement.
+  - feature: pre-compile the pattern to improve the performance.